Separated the CommandHandler
Divided the client's CommandHandler class into partial classes.
