+++ Client/Client.csproj
@@ -47,6 +47,11 @@
   <ItemGroup>
     <Compile Include="Core\Client.cs" />
     <Compile Include="Core\Commands\CommandHandler.cs" />
+    <Compile Include="Core\Commands\ConnectionHandler.cs" />
+    <Compile Include="Core\Commands\FileHandler.cs" />
+    <Compile Include="Core\Commands\MiscHandler.cs" />
+    <Compile Include="Core\Commands\SurveillanceHandler.cs" />
+    <Compile Include="Core\Commands\SystemHandler.cs" />
     <Compile Include="Core\Compression\SafeQuickLZ.cs" />
     <Compile Include="Core\Elevation\CommandButton.cs">
       <SubType>Component</SubType>
